requirements.txt
imagehash
elasticsearch
Pillow